+++ datacontract/cli.py
@@ -67,7 +67,7 @@ def common(
     pass
 
 
-@app.command()
+@app.command(name="init")
 def init(
     location: Annotated[
         str, typer.Argument(help="The location of the data contract file to create.")
@@ -90,7 +90,7 @@ def init(
     console.print("📄 data contract written to " + location)
 
 
-@app.command()
+@app.command(name="lint")
 def lint(
     location: Annotated[
         str,
@@ -125,7 +125,7 @@ def enable_debug_logging(debug: bool):
         )
 
 
-@app.command()
+@app.command(name="test")
 def test(
     location: Annotated[
         str,
@@ -187,7 +187,7 @@ def test(
     write_test_result(run, console, output_format, output, data_contract)
 
 
-@app.command()
+@app.command(name="export")
 def export(
     format: Annotated[ExportFormat, typer.Option(help="The export format.")],
     output: Annotated[
@@ -199,10 +199,7 @@ def export(
     server: Annotated[str, typer.Option(help="The server name to export.")] = None,
     schema_name: Annotated[
         str,
-        typer.Option(
-            help="The name of the schema to export, e.g., `orders`, or `all` for all "
-            "schemas (default)."
-        ),
+        typer.Option(help="The name of the schema to export, e.g., `orders`, or `all` for all schemas (default)."),
     ] = "all",
     # TODO: this should be a subcommand
     rdf_base: Annotated[
@@ -431,7 +428,11 @@ def catalog(
     enable_debug_logging(debug)
 
     path = Path(output)
-    path.mkdir(parents=True, exist_ok=True)
+    try:
+        path.mkdir(parents=True, exist_ok=True)
+    except FileExistsError:
+        if not path.is_dir():
+            raise
     console.print(f"Created {output}")
 
     contracts = []
@@ -467,7 +468,7 @@ def _get_uvicorn_arguments(port: int, host: str, context: typer.Context) -> dict
     return default_args | dict(zip(trimmed_keys, context.args[1::2]))
 
 
-@app.command(context_settings={"allow_extra_args": True, "ignore_unknown_options": True})
+@app.command(name="api", context_settings={"allow_extra_args": True, "ignore_unknown_options": True})
 def api(
     ctx: Annotated[typer.Context, typer.Option(help="Extra arguments to pass to uvicorn.run().")],
     port: Annotated[int, typer.Option(help="Bind socket to this port.")] = 4242,