+++ datacontract/imports/parquet_importer.py
@@ -14,9 +14,7 @@ from datacontract.model.exceptions import DataContractException
 
 
 class ParquetImporter(Importer):
-    def import_source(
-        self, source: str, import_args: dict
-    ) -> OpenDataContractStandard:
+    def import_source(self, source: str, import_args: dict) -> OpenDataContractStandard:
         return import_parquet(source)